requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
Pillow==10.1.0
seaborn==0.13.2
requests==2.31.0
streamlit==1.32.2 